Allow default tests to be overwritten with parametrize
