+++ src/stratocaster/tests/test_connectivity_strategy.py
@@ -23,63 +23,15 @@ class TestConnectivityStrategy(StrategyTestMixin):
     valid_settings = {(0.5, 0.1, 10), (0.1, None, 10), (0.5, 0.1, None)}
 
     @pytest.mark.parametrize(["decay_rate", "cutoff", "max_runs"], valid_settings)
-    def test_simulated_termination(
-        self, benzene_variants_star_map, decay_rate, cutoff, max_runs
-    ):
+    def test_simulated_termination(self, fanning_network, decay_rate, cutoff, max_runs):
 
         settings = ConnectivityStrategySettings(
             decay_rate=decay_rate, cutoff=cutoff, max_runs=max_runs
         )
-        default_strategy = self.strategy_class(settings)
 
-        def counts_to_result_data(counts_dict):
-            result_data = {}
-            for transformation_key, count in counts_dict.items():
-                result = DummyProtocolResult(
-                    n_protocol_dag_results=count, info=f"key: {transformation_key}"
-                )
-                result_data[transformation_key] = result
-            return result_data
-
-        def shuffle_take_n(keys_list, n):
-            shuffle(keys_list)
-            return keys_list[:n]
-
-        # initial transforms
-        transformation_counts = {
-            transformation.key: 0 for transformation in benzene_variants_star_map.edges
-        }
-
-        max_iterations = 100
-        current_iteration = 0
-        while current_iteration <= max_iterations:
-
-            if current_iteration == max_iterations:
-                raise RuntimeError(
-                    f"Strategy did not terminate in {max_iterations} iterations "
-                )
-
-            result_data = counts_to_result_data(transformation_counts)
-            proposal = default_strategy.propose(benzene_variants_star_map, result_data)
-
-            # get random transformations from those with a non-None weight
-            resolved_keys = shuffle_take_n(
-                [
-                    key
-                    for key, weight in proposal.resolve().items()
-                    if weight is not None
-                ],
-                5,
-            )
-
-            if resolved_keys:
-                # pretend we ran each of the randomly selected protocols
-                for key in resolved_keys:
-                    transformation_counts[key] += 1
-            # if we got an empty list back, there are not more protocols to run
-            else:
-                break
-            current_iteration += 1
+        StrategyTestMixin.test_simulated_termination(
+            self, fanning_network, settings=settings
+        )
 
     def test_propose_cutoff_num_runs_predictioned_termination(
         self, benzene_variants_star_map
@@ -90,9 +42,7 @@ class TestConnectivityStrategy(StrategyTestMixin):
         """
 
         settings = ConnectivityStrategySettings(cutoff=2, decay_rate=0.5)
-        strategy = ConnectivityStrategy(settings)
-
-        assert isinstance(settings.cutoff, float)
+        strategy = self.strategy_or_default(settings)
 
         num_runs = math.floor(
             math.log(settings.cutoff / 3.5) / math.log(settings.decay_rate)
